feat: ヘッダーメニューにログアウトを追加 #139
メニューをまとめるため
