+++ frontend/src/components/HeaderMenuButton.tsx
@@ -1,11 +1,17 @@
 import { FC, MouseEvent, useContext, useState } from "react";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
+import Cookies from "js-cookie";
 import { Avatar, Box, Button, Divider, Menu, MenuItem } from "@mui/material";
 import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
+import { AxiosRequestConfig, AxiosResponse } from "axios";
+import { axiosInstance } from "../utils/axios";
 import { AuthContext } from "../providers/AuthProvider";
+import { SnackbarContext } from "../providers/SnackbarProvider";
 
 export const HeaderMenuButton: FC = () => {
-  const { currentUser } = useContext(AuthContext);
+  const { setIsSignedIn, currentUser } = useContext(AuthContext);
+  const { handleSetSnackbar } = useContext(SnackbarContext);
+  const navigate = useNavigate();
 
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const open = Boolean(anchorEl);
@@ -16,6 +22,45 @@ export const HeaderMenuButton: FC = () => {
 
   const handleClose = () => {
     setAnchorEl(null);
+  };
+
+  const handleSignOut = () => {
+    const options: AxiosRequestConfig = {
+      url: "/auth/sign_out",
+      method: "DELETE",
+      headers: {
+        "access-token": Cookies.get("_access_token") || "",
+        client: Cookies.get("_client") || "",
+        uid: Cookies.get("_uid") || "",
+      },
+    };
+
+    axiosInstance(options)
+      .then((res: AxiosResponse) => {
+        console.log(res);
+
+        if (res.status === 200) {
+          Cookies.remove("_access_token");
+          Cookies.remove("_client");
+          Cookies.remove("_uid");
+          setIsSignedIn(false);
+          handleSetSnackbar({
+            open: true,
+            type: "success",
+            message: "ログアウトしました",
+          });
+          navigate("/", { replace: true });
+        }
+      })
+      .catch((err) => {
+        console.log(err);
+        handleSetSnackbar({
+          open: true,
+          type: "error",
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+          message: `${err.response.data.errors[0]}`,
+        });
+      });
   };
 
   return (
@@ -68,6 +113,7 @@ export const HeaderMenuButton: FC = () => {
         >
           ユーザー設定
         </MenuItem>
+        <MenuItem onClick={handleSignOut}>ログアウト</MenuItem>
       </Menu>
     </Box>
   );
